moving formatting to format folder
