+++ ns_bot/controllers/nation_controller.py
@@ -7,7 +7,7 @@ from discord.ext import commands
 
 from ns_bot.controllers.base_nationstate_controller import BaseNationstateController
 from ns_bot.data.shards import VALID_PUBLIC_NATION_SHARDS
-from ns_bot.utils.formatter import format_nation_info
+from ns_bot.formatting.nation_info import format_nation_info
 from ns_bot.utils.wrappers import async_wrapper
