+++ ns_bot/controllers/region_controller.py
@@ -7,7 +7,7 @@ from discord.ext import commands
 
 from ns_bot.controllers.base_nationstate_controller import BaseNationstateController
 from ns_bot.data.shards import VALID_REGION_SHARDS
-from ns_bot.utils.formatter import format_region_info
+from ns_bot.formatting.nation_info import format_region_info
 
 
 class RegionController(BaseNationstateController):
